+++ src/hooks/index.ts
@@ -2,5 +2,4 @@ export { default as useFadeNavigate } from './useFadeNavigate';
 export { default as useSelectBox } from './useSelectBox';
 export { default as useSignupForm } from './useSignupForm';
 export { default as useCustomToast } from './useCustomToast';
-export { default as useDeleteAccountForm } from './useDeleteAccountForm';
 export { default as useUserInfoForm } from './useUserInfoForm';
